+++ src/components/Account/Influencer/SignupSocial/index.tsx
@@ -56,16 +56,21 @@ const SignupSocialInfluencer = () => {
             입력해주세요.{" "}
             <span className={styles["text-require"]}>&#40;필수&#41;</span>
           </p>
-          <SNSInput id="blog" type="blog" placeholder="네이버 블로그" gap={5} />
           <SNSInput
-            id="instargram"
-            type="instargram"
+            id="blog"
+            type="NAVER_BLOG"
+            placeholder="네이버 블로그"
+            gap={5}
+          />
+          <SNSInput
+            id="instagram"
+            type="INSTAGRAM"
             placeholder="인스타그램"
             gap={5}
           />
-          <SNSInput id="youtube" type="youtube" placeholder="유튜브" gap={5} />
-          <SNSInput id="tictok" type="tictok" placeholder="틱톡" gap={5} />
-          <SNSInput id="etc" type="etc" placeholder="기타" gap={5} />
+          <SNSInput id="youtube" type="YOUTUBE" placeholder="유튜브" gap={5} />
+          <SNSInput id="tictok" type="TIKTOK" placeholder="틱톡" gap={5} />
+          <SNSInput id="etc" type="ETC" placeholder="기타" gap={5} />
         </div>
         <TermsCheck />
       </form>
