+++ src/components/Mypage/Employer/CampaignItem/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Image from "next/image";
 import ms from "@/utils/modifierSelector";
-import IconInsta from "@/assets/icons/icon-sns-instargram.svg";
+import IconInsta from "@/assets/icons/icon-sns-instagram.svg";
 import IconHeartWhite from "@/assets/icons/icon-heart-white.svg";
 import IconPointCoin from "@/assets/icons/icon-point-coin.svg";
 import Tag from "@/components/Tag";
